[MOD] moved sidebar into components
