+++ src/Users/Admin/AdmAccountsPage.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import HeaderComp from '../../Components/Header/HeaderComp'
-import Sidebar from '../../Components/AdminMenu/index'
+import Sidebar from '../../Components/Sidebar/Sidebar'
 import AdmAccountsTable from '../../Components/Accounts/Adm/AdmAccountsTable.js'
 import AdmAccountCreation from '../../Components/Accounts/Adm/AdmAccountCreation.js'
 import './AdmAccountsPage.css'
