+++ src/Users/Admin/AdmHomePage.jsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { LastAlerts } from '../../Components/Alerts/LastAlerts'
 import HeaderComp from '../../Components/Header/HeaderComp'
-import Sidebar from '../../Components/AdminMenu/index'
+import Sidebar from '../../Components/Sidebar/Sidebar'
 import './AdmHomePage.css'
 
 export default function AdmHomePage () {
