+++ src/index.jsx
@@ -1,7 +1,6 @@
 import React from 'react'
 import { createRoot } from 'react-dom/client'
 import { BrowserRouter as Router, Routes, Route, Navigate } from 'react-router-dom'
-import './styles/sidebar.scss'
 import Login from './Users/Public/loginPage'
 import AdmHomePage from './Users/Admin/AdmHomePage'
 import AdmAccountsPage from './Users/Admin/AdmAccountsPage'
@@ -28,21 +27,15 @@ if (rootElement) {
           </>
         )}
         {sessionStorage.getItem('role') !== null && (
-          <Route
-            path='/login'
-            element={<Navigate to='/' replace />}
-          />,
-          <Route
-            path='/forgot'
-            element={<Navigate to='/' replace />}
-          />
+          <Route path='/login' element={<Navigate to='/' replace />} />,
+          <Route path='/forgot' element={<Navigate to='/' replace />} />
         )}
-        {sessionStorage.getItem('role') === 'admin' && (
+        {/* {sessionStorage.getItem('role') === 'admin' && (
           <>
             <Route path='/' element={<AdmHomePage />} />
-            <Route exact path='/accounts' element={<AdmAccountsPage />} />
+            <Route path='/accounts' element={<AdmAccountsPage />} />
           </>
-        )}
+        )} */}
         {sessionStorage.getItem('role') === 'administration' && (
           <>
             <Route path='/' element={<SchoolAdmHomePage />} />
